+++ vitest.config.ts
@@ -12,7 +12,7 @@ export default defineConfig({
       extends: true,
       test: {
         environment: 'node',
-        include: ['src/**/*.test.ts']
+        include: ['src/**/*.test.{ts,tsx}']
       }
     }, {
       extends: true,